refactor: Remove unnecessary code and organize files into folders
- 기능 수정 없음
