+++ app/src/main/java/campus/tech/kakao/map/activity/MapErrorActivity.kt
@@ -1,12 +1,12 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.activity
 
 import android.os.Bundle
-import android.util.Log
 import android.widget.TextView
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import campus.tech.kakao.map.R
 
 class MapErrorActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
