+++ app/src/main/java/campus/tech/kakao/map/adapter/SearchDataAdapter.kt
@@ -1,15 +1,12 @@
 package campus.tech.kakao.map.adapter
 
-import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.RecentAdapterListener
-import campus.tech.kakao.map.SearchAdapterListener
-import campus.tech.kakao.map.activity.HomeMapActivity
+import campus.tech.kakao.map.listener.SearchAdapterListener
 import campus.tech.kakao.map.viewModel.RecentViewModel
 import campus.tech.kakao.map.retrofit.Document
