+++ app/src/main/java/campus/tech/kakao/map/listener/RecentAdapterListener.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.listener
 
 interface RecentAdapterListener {
     fun autoSearch(searchData: String)
