+++ app/src/main/java/campus/tech/kakao/map/listener/SearchAdapterListener.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.listener
 
 interface SearchAdapterListener {
     fun displaySearchLocation(name: String, address: String, latitude: String, longitude: String)
