+++ app/src/main/res/layout/activity_map_error.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapErrorActivity">
+    tools:context=".activity.MapErrorActivity">
 
     <TextView
         android:id="@+id/error_notice"
